updated code

+++ source/code-snippets/usage-examples/changeStreamsTwo.js
@@ -1,12 +1,10 @@
-// ignored first line
 const { MongoClient } = require("mongodb");
-
-// Replace the following with your MongoDB deployment's connection string.
+// Replace the following with your MongoDB deployment's connection
+// string.
 const uri =
-  "mongodb+srv://<user>:<password>@<cluster-url>?retryWrites=true&w=majority";
-
+  "mongodb+srv://admin:admin@mflix-2sp0m.mongodb.net/test?retryWrites=true&w=majority";
 const client = new MongoClient(uri, { useUnifiedTopology: true });
-let changeStream;
+let changeStream, interval;
 async function run() {
   try {
     await client.connect();
@@ -19,16 +17,21 @@ async function run() {
       // process a change event
       console.log("a change to the collection happened: \t", next);
     });
-    setTimeout(async () => {
-      console.log("inserting");
-      await collection.insertOne({ test: "change stream" });
-      setTimeout(async () => {
+    let actions = 0;
+    interval = setInterval(async () => {
+      if (actions === 0) {
+        actions++;
+        console.log("inserting");
+        return await collection.insertOne({ test: "change stream" });
+      }
+      if (actions === 1) {
         console.log("deleting");
-        await collection.deleteOne({ test: "change stream" });
-      });
+        return await collection.deleteOne({ test: "change stream" });
+      }
     }, 1000);
   } finally {
     setTimeout(async () => {
+      clearInterval(interval);
       await changeStream.close();
       await client.close();
     }, 2500);
